some code reorg
