+++ GW2EIEvtcParser/EncounterLogic/Strikes/IBS/Bjora/FraenirOfJormag.cs
