+++ GW2EIEvtcParser/EncounterLogic/Strikes/IBS/Bjora/SuperKodanBrothers.cs
