+++ GW2EIEvtcParser/EncounterLogic/Strikes/IBS/Bjora/WhisperOfJormag.cs
