+++ GW2EIEvtcParser/EncounterLogic/Strikes/IBS/Drizzlewood/ColdWar.cs
@@ -5,7 +5,7 @@ using GW2EIEvtcParser.Exceptions;
 
 namespace GW2EIEvtcParser.EncounterLogic
 {
-    internal class ColdWar : StrikeMissionLogic
+    internal class ColdWar : IBSStrike
     {
         public ColdWar(int triggerID) : base(triggerID)
         {
